Developing layout & responsive behavior of quick-tabs
Designed a crimson and gray themed layout for a quick-tabs feature for which the tabs and assocaited panels are easily edited using the WordPress interface with minimal need for direct HTML coding. Also designed the JavaScript code that drives activation & deactivation of panels by interaction with tabs, as well as automated scrolling behavior for changes in document height associated with changes in active panel height.